requirements.txt
discord.py-self
requests
psycopg==3.1.4
psycopg-binary==3.1.4